requirements.txt
asgiref==3.7.2
Django==4.2.3
sqlparse==0.4.4
typing_extensions==4.7.1
Pillow==10.0.0
djangorestframework==3.14.0
django-rest-auth==0.9.5
djangorestframework_simplejwt==5.2.2
django_bootstrap5
django-allauth==0.54.0
djangorestframework-api-key
django-bootstrap4
django-avatar
django-countries